requirements.txt
django
dj-database-url
djangorestframework
pillow
psycopg2-binary
django-cors-headers
djangorestframework_simplejwt
WeasyPrint
drf-api-logger
gunicorn
python-slugify